Move Operator to its own module
